Add and Implement functions
Added and Implement functions in BaseballGameManager.

* showMenu()
This function prints a menu that the user can select.

The reason why this function was added and implemented was
to add and implement to satisfy the input example (입력예시)
of Requirement 1 of step-2.

* receiveUserMenuSelect()
This function receives input by user.

The reason why this function was added and implemented was
to add and implement to satisfy the input example (입력예시)
of Requirement 1 of step-2.

* performUserMenuSelection()
This function allows caller to perform each result on user input.